Refaktoroi: Splittaa full ja inkrementaalinen eri luokkiin
